refactor: Move limiter to debug
