+++ compat/test/browser/render.test.js
@@ -3,9 +3,7 @@ import React, {
 	render,
 	Component,
 	hydrate,
-	createContext,
-	useState,
-	Fragment
+	createContext
 } from 'preact/compat';
 import { setupRerender, act } from 'preact/test-utils';
 import {
@@ -564,36 +562,5 @@ describe('compat render', () => {
 		React.render(<Foo />, scratch);
 
 		expect(scratch.textContent).to.equal('foo');
-	});
-
-	it('throws an error if a component rerenders too many times', () => {
-		let rerenderCount = 0;
-		function TestComponent({ loop = false }) {
-			const [count, setCount] = useState(0);
-			if (loop) {
-				setCount(count + 1);
-			}
-
-			if (count > 30) {
-				expect.fail(
-					'Repeated rerenders did not cause the expected error. This test is failing.'
-				);
-			}
-
-			rerenderCount += 1;
-			return <div />;
-		}
-
-		expect(() => {
-			render(
-				<Fragment>
-					<TestComponent />
-					<TestComponent loop />
-				</Fragment>,
-				scratch
-			);
-		}).to.throw(/Too many re-renders/);
-		// 1 for first TestComponent + 24 for second TestComponent
-		expect(rerenderCount).to.equal(25);
 	});
 });
